Create check_ping.sh
A regular check will be performed to verify if WireGuard is active by pinging an internal device on IP address 192.168.1.254. If WireGuard is already active, no action will be taken. In the event that WireGuard is not running, an attempt will be made to start it. If the ping is unsuccessful, the DNS will be flushed to account for any changes in the domain's IP address, and WireGuard will be restarted. This check will be scheduled to run every minute using cron.